requirements.txt
django
djangorestframework
pillow
drf-yasg
celery
django-filter
djangorestframework-simplejwt

 
